% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date11.R
\name{date11}
\alias{date11}
\title{Convert date to Date11. format}
\usage{
date11(date, informat)
}
\arguments{
\item{date}{Input date}

\item{informat}{Input date format}
}
\value{
Date to Date11. format
}
\description{
Convert date to Date11. format
}
\examples{
date11(date = "19-10-1994",informat = "\%d-\%m-\%Y")

}
