% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date9.R
\name{date9}
\alias{date9}
\title{Convert date to Date9. format}
\usage{
date9(date, informat)
}
\arguments{
\item{date}{Input date}

\item{informat}{Input date format}
}
\value{
Date to Date9. format\
}
\description{
Convert date to Date9. format
}
\examples{
 date9(date = "19-10-1994",informat = "\%d-\%m-\%Y")
}
