% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddmmyy10.R
\name{ddmmyy10}
\alias{ddmmyy10}
\title{Convert date to ddmmyy10. format}
\usage{
ddmmyy10(date, informat)
}
\arguments{
\item{date}{Input date}

\item{informat}{Input date format}
}
\value{
Date to ddmmyy10. format
}
\description{
Convert date to ddmmyy10. format
}
\examples{
ddmmyy10(date = "19-10-1994",informat = "\%d-\%m-\%Y")
}
