% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso8601da.R
\name{iso8601da}
\alias{iso8601da}
\title{Convert date to ISO8601 format}
\usage{
iso8601da(date, informat)
}
\arguments{
\item{date}{Input date}

\item{informat}{Input date format}
}
\value{
Date to ISO8601da. format
}
\description{
Convert date to ISO8601 format
}
\examples{
iso8601da(date = "19-10-1994",informat = "\%d-\%m-\%Y")
}
