% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso8601dt.R
\name{iso8601dt}
\alias{iso8601dt}
\title{Convert date and time to ISO8601 format}
\usage{
iso8601dt(date, time, informat)
}
\arguments{
\item{date}{Input date}

\item{time}{Input time}

\item{informat}{Input date format}
}
\value{
Date and time in ISO8601 format
}
\description{
Convert date and time to ISO8601 format
}
\examples{
iso8601dt(date = "19-10-1994",time = "10:10",informat = "\%Y-\%m-\%d \%H:\%M")
}
