% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_date.R
\name{num_date}
\alias{num_date}
\title{Convert SAS Numeric date values to Date}
\usage{
num_date(date)
}
\arguments{
\item{date}{numeric date}
}
\value{
SAS date
}
\description{
Convert SAS Numeric date values to Date
}
\examples{
num_date(20562)

}
